requirements.txt
lmdb
psutil
loguru
pillow
numpy
pandas
pyyaml